+++ CHANGELOG.md
@@ -9,6 +9,12 @@ This Changelog tracks changes to this project. The notes below include a summary
 - `fixed` for any bug fixes.
 - `security` in case of vulnerabilities.
 
+## Version `0.26.3` - 19 Oct 2022
+
+- `changed` pipeline configuration generated for RIMA
+- `added` ingestion sheet generation for RIMA
+- `removed` metasheet generation for RIMA
+
 ## Version `0.26.2` - 10 Oct 2022
 
 - `added` MIBI assay support